functions/webhook/requirements.txt
parliament-functions==0.1.0
cloudevents==1.11.0
functions-framework==3.*
flask==3.1.0
minio==7.2.15
requests==2.32.0